JAVA-45628 Created New Module spring-ai-3 & Moved code of article spring-ai-mongodb-rag
