+++ spring-ai-3/README.md
@@ -1,0 +1,2 @@
+## Relevant Articles
+- [Building a RAG App Using MongoDB and Spring AI](https://www.baeldung.com/spring-ai-mongodb-rag)
