+++ spring-ai-3/src/main/java/com/baeldung/springai/rag/mongodb/configuration/AdvisorConfiguration.java
@@ -11,6 +11,6 @@ public class AdvisorConfiguration {
 
     @Bean
     public QuestionAnswerAdvisor questionAnswerAdvisor(VectorStore vectorStore) {
-        return new QuestionAnswerAdvisor(vectorStore, SearchRequest.defaults());
+        return new QuestionAnswerAdvisor(vectorStore, SearchRequest.builder().build());
     }
 }
