+++ spring-ai-3/src/main/java/com/baeldung/springai/rag/mongodb/dto/WikiDocument.java
