+++ spring-ai-3/src/test/docker/mongodb/docker-compose.yml
