+++ spring-ai-3/src/test/resources/application.yml
